+++ plugins/modules/na_ontap_cluster_peer.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2018-2025, NetApp, Inc
+# (c) 2018-2026, NetApp, Inc
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 from __future__ import absolute_import, division, print_function
 __metaclass__ = type
@@ -98,12 +98,34 @@ options:
       - By default the system will generate the same name as cluster name.
     type: str
     version_added: '23.1.0'
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: '23.4.0'
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
 short_description: NetApp ONTAP Manage Cluster peering
 version_added: 2.7.0
 
 notes:
   - Modify remote intercluster addresses operation is supported only with REST.
   - The options local_name_for_peer and local_name_for_source are supported only with REST.
+  - Supports AWS Lambda proxy functionality when using REST. See the README file for example usage.
 '''
 
 EXAMPLES = """
@@ -199,7 +221,9 @@ class NetAppONTAPClusterPeer:
             local_name_for_peer=dict(required=False, type='str'),
             local_name_for_source=dict(required=False, type='str'),
         ))
-
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
+        # Add Lambda support to peer_options
+        self.argument_spec['peer_options']['options'].update(netapp_utils.na_ontap_lambda_argument_spec())
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
             mutually_exclusive=[
@@ -210,7 +234,8 @@ class NetAppONTAPClusterPeer:
             required_one_of=[['peer_options', 'dest_hostname']],
             required_if=[
                 ('state', 'absent', ['source_cluster_name', 'dest_cluster_name']),
-                ('state', 'present', ['source_intercluster_lifs', 'dest_intercluster_lifs'])
+                ('state', 'present', ['source_intercluster_lifs', 'dest_intercluster_lifs']),
+                ('use_lambda', True, ['lambda_config',])
             ],
             supports_check_mode=True
         )
@@ -235,6 +260,8 @@ class NetAppONTAPClusterPeer:
         self.dst_use_rest = self.dst_rest_api.is_rest()
         self.use_rest = bool(self.src_use_rest and self.dst_use_rest)
         if not self.use_rest:
+            if self.parameters.get('use_lambda'):
+                self.module.fail_json(msg="Error: AWS Lambda proxy for ONTAP APIs is only supported with REST.")
             if not netapp_utils.has_netapp_lib():
                 self.module.fail_json(msg="the python NetApp-Lib module is required")
             self.server = netapp_utils.setup_na_ontap_zapi(module=self.module)